+++ test/mock/node/lib/app.js
@@ -14,11 +14,25 @@ function bar(argument) {
     }
 }
 
+const a = () => {};
+
+(function() {})();
 
 function app() {
     console.log('this is app');
 }
 
+if (module.a) {
+    console.log('a');
+}
+
+for (let i = 0; i < 5; i++) {
+    if (i > 2) {
+        console.log(i);
+    } else if (module.b) {
+        console.log('b');
+    }
+}
 
 module.exports = {
     app,