+++ test/mock/src/async.js
@@ -1,0 +1,24 @@
+const f = false;
+const t = true;
+console.info('i am a line of code');
+
+function apple(awesome) {
+    if (f || t) {
+        console.info('what');
+    }
+    if (t || f) {
+        console.log('hey');
+    }
+}
+
+function missed() {
+
+}
+
+function missed2() {
+
+}
+
+apple();
+apple();
+apple();