+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Changelog
 
+* 2.4.0
+    - (New Feature) added statements to report
+    - optimized source mapping related codes
+
 * 2.3.4
     - fixed ignore lines
     - fixed AssignmentPattern branches